refactor: unify common range

+++ specs/src/configure_table.rs
@@ -2,12 +2,12 @@ use serde::Serialize;
 
 pub const WASM_PAGE_SIZE: u64 = 65536;
 
-const WASM_32_MAXIMAL_PAGES_DEFAULT: usize = 65536;
+const WASM_32_MAXIMAL_PAGES_DEFAULT: u32 = 65536;
 
 #[derive(Serialize, Debug, Clone, Copy)]
 pub struct ConfigureTable {
-    pub init_memory_pages: usize,
-    pub maximal_memory_pages: usize,
+    pub init_memory_pages: u32,
+    pub maximal_memory_pages: u32,
 }
 
 impl Default for ConfigureTable {
