+++ specs/src/imtable.rs
@@ -5,7 +5,7 @@ use serde::Serialize;
 pub struct InitMemoryTableEntry {
     pub ltype: LocationType,
     pub is_mutable: bool,
-    pub offset: u64,
+    pub offset: u32,
     pub vtype: VarType,
     /// convert from [u8; 8] via u64::from_le_bytes
     pub value: u64,
@@ -29,7 +29,7 @@ impl InitMemoryTable {
         serde_json::to_string(&self.0).unwrap()
     }
 
-    pub fn try_find(&self, ltype: LocationType, offset: u64) -> Option<u64> {
+    pub fn try_find(&self, ltype: LocationType, offset: u32) -> Option<u64> {
         for entry in self.0.iter() {
             if entry.ltype == ltype && entry.offset == offset {
                 return Some(entry.value);
@@ -47,7 +47,7 @@ impl InitMemoryTable {
         self.0.iter().filter(|e| e.ltype == ltype).collect()
     }
 
-    pub fn first_consecutive_zero_memory(&self) -> u64 {
+    pub fn first_consecutive_zero_memory(&self) -> u32 {
         self.0.last().map_or(0, |entry| entry.offset + 1)
     }
 }
